Изменение структуры программы. Исключение части программы.
